requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.20
langchain-openai==0.1.8
langchain-community==0.0.38
langchain-core==0.1.52
chromadb==0.4.22
sentence-transformers==2.2.2
pypdf2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4